XML value configuration and disable date which not include in selected month.

+++ CalendarView/src/main/java/com/shahzadafridi/calendarview/DayConfiguration.kt
@@ -6,6 +6,6 @@ data class DayConfiguration(
     var daySelectedClr: Int? = null,
     var daySelectedBg: Int? = null,
     var dayFont: Int? = null,
-    var dayTxtSize: Int? = null,
+    var dayTxtSize: Float? = null,
     var eventDotColor: Int? = null
 )
